+++ pkg/exporter/sumologicexporter/config_test.go
@@ -55,6 +55,19 @@ func TestInitExporterInvalidConfiguration(t *testing.T) {
 			},
 		},
 		{
+			name:          "unsupported Graphite metrics format",
+			expectedError: errors.New("support for the graphite metric format was removed, please use prometheus or otlp instead"),
+			cfg: &Config{
+				LogFormat:    "json",
+				MetricFormat: "graphite",
+				HTTPClientSettings: confighttp.HTTPClientSettings{
+					Timeout:  defaultTimeout,
+					Endpoint: "test_endpoint",
+				},
+				CompressEncoding: "gzip",
+			},
+		},
+		{
 			name:          "unexpected trace format",
 			expectedError: errors.New("unexpected trace format: text"),
 			cfg: &Config{