+++ pkg/exporter/sumologicexporter/sender.go
@@ -136,7 +136,6 @@ type sender struct {
 	sources             sourceFormats
 	compressor          compressor
 	prometheusFormatter prometheusFormatter
-	graphiteFormatter   graphiteFormatter
 	jsonLogsConfig      JSONLogs
 	dataUrlMetrics      string
 	dataUrlLogs         string
@@ -161,7 +160,6 @@ const (
 
 	contentTypeLogs       string = "application/x-www-form-urlencoded"
 	contentTypePrometheus string = "application/vnd.sumologic.prometheus"
-	contentTypeGraphite   string = "application/vnd.sumologic.graphite"
 	contentTypeOTLP       string = "application/x-protobuf"
 
 	contentEncodingGzip    string = "gzip"
@@ -175,7 +173,6 @@ func newSender(
 	s sourceFormats,
 	c compressor,
 	pf prometheusFormatter,
-	gf graphiteFormatter,
 	metricsUrl string,
 	logsUrl string,
 	tracesUrl string,
@@ -187,7 +184,6 @@ func newSender(
 		sources:             s,
 		compressor:          c,
 		prometheusFormatter: pf,
-		graphiteFormatter:   gf,
 		jsonLogsConfig:      cfg.JSONLogs,
 		dataUrlMetrics:      metricsUrl,
 		dataUrlLogs:         logsUrl,
@@ -554,8 +550,6 @@ func (s *sender) sendNonOTLPMetrics(ctx context.Context, md pmetric.Metrics) (pm
 				switch s.config.MetricFormat {
 				case PrometheusFormat:
 					formattedLine = s.prometheusFormatter.metric2String(m, rm.Resource().Attributes())
-				case GraphiteFormat:
-					formattedLine = s.graphiteFormatter.metric2String(m, rm.Resource().Attributes())
 				default:
 					return md, []error{fmt.Errorf("unexpected metric format: %s", s.config.MetricFormat)}
 				}
@@ -744,8 +738,6 @@ func addMetricsHeaders(req *http.Request, mf MetricFormatType) error {
 	switch mf {
 	case PrometheusFormat:
 		req.Header.Add(headerContentType, contentTypePrometheus)
-	case GraphiteFormat:
-		req.Header.Add(headerContentType, contentTypeGraphite)
 	case OTLPMetricFormat:
 		req.Header.Add(headerContentType, contentTypeOTLP)
 	default: